同步 DuiDesigner 至新版 DuiLib
统一 VS2026、x64、Unicode、C++20 和 MT/MTd 配置，补齐当前控件映射并修复 x64 兼容问题。

+++ DuiDesigner/DialogUIAttribEdit.cpp
@@ -1,11 +1,11 @@
-// DialogUIAttribEdit.cpp : ÊµÏÖÎÄ¼þ
+// DialogUIAttribEdit.cpp : å®žçŽ°æ–‡ä»¶
 //
 #include "stdafx.h"
 #include "DuiDesigner.h"
 #include "DialogUIAttribEdit.h"
 
 
-// CDialogUIAttribEdit ¶Ô»°¿ò
+// CDialogUIAttribEdit å¯¹è¯æ¡†
 
 IMPLEMENT_DYNAMIC(CDialogUIAttribEdit, CDialog)
 
@@ -36,13 +36,13 @@ BEGIN_MESSAGE_MAP(CDialogUIAttribEdit, CDialog)
 END_MESSAGE_MAP()
 
 
-// CDialogUIAttribEdit ÏûÏ¢´¦Àí³ÌÐò
+// CDialogUIAttribEdit æ¶ˆæ¯å¤„ç†ç¨‹åº
 
 BOOL CDialogUIAttribEdit::OnInitDialog()
 {
 	CDialog::OnInitDialog();
 
-	// TODO:  ÔÚ´ËÌí¼Ó¶îÍâµÄ³õÊ¼»¯
+	// TODO:  åœ¨æ­¤æ·»åŠ é¢å¤–çš„åˆå§‹åŒ–
 	CRect rectUIProp;
 	m_wndUIPropLocation.GetClientRect(&rectUIProp);
 	m_wndUIPropLocation.MapWindowPoints(this, &rectUIProp);
@@ -50,13 +50,13 @@ BOOL CDialogUIAttribEdit::OnInitDialog()
 	m_wndUIProperties.ShowProperty(m_pControl);
 
 	return TRUE;  // return TRUE unless you set the focus to a control
-	// Òì³£: OCX ÊôÐÔÒ³Ó¦·µ»Ø FALSE
+	// å¼‚å¸¸: OCX å±žæ€§é¡µåº”è¿”å›ž FALSE
 }
 
 LRESULT CDialogUIAttribEdit::OnUIPropChanged(WPARAM wp, LPARAM lp)
 {
 	CMFCPropertyGridProperty* pProp = (CMFCPropertyGridProperty *)lp;
-	SetUIValue(pProp, pProp->GetData());
+	SetUIValue(pProp, static_cast<int>(pProp->GetData()));
 
 	return TRUE;
 }
